ultimo commit antes del cambio de version 5

+++ app/src/main/java/com/example/kramviapp/models/CreateProformaModel.kt
@@ -5,7 +5,7 @@ import com.example.kramviapp.enums.InvoiceType
 
 data class CreateProformaModel(
     val discount: Double,
-    val igvPercent: Int,
+    val igvPercent: Double,
     val currencyCode: CurrencyCodeType,
     val observations: String,
     val customerId: String?,
